Removed redundent type cast

+++ CircularLayout/src/org/gephi/layout/plugin/circularlayout/CircleLayout.java
@@ -94,8 +94,8 @@ public class CircleLayout extends AbstractLayout implements Layout {
                    Arrays.sort(nodes, new Comparator<Node>() {
                         @Override
                         public int compare(Node o1, Node o2) {
-                            int f1 = (int) graph.getDegree(o1);
-                            int f2 = (int) graph.getDegree(o2);
+                            int f1 = graph.getDegree(o1);
+                            int f2 = graph.getDegree(o2);
                           return f2-f1;
                         }
                     });
